+++ libraries/OpenHive/OpenHive.h
@@ -1,0 +1,2 @@
+#include "OpenHiveSensors.h"
+#include "OpenHiveTemperatureArray.h"